This commit seems to be working as I think it is supposed to. I will walk through code with Daniel in the morning to verify.
Now custom voicing generator and hand manager are pulled into Style.
When ChordPattern needs to use this, they are obtained from the Style in which the pattern is rendered.
The voicing parameters are loaded when the Style is created, not when each chord is voiced.

Wholesale getters were added to move parameters into the voicing generator and hand manager, rather than doing this piecemeal as before. These getters should be reviewd to make sure they are getting all the essentials.

There is no more use of the default voicings in ImproVisor.java as before. 

Positioning of various methods such as resetHands() should be reviewed.