Here's the reason the standard voicings don't return after using a custom voicing:
This method:

public static Polylist getVoicingAndExtensionList(ChordSymbol chord,
                                                  Polylist lastChord,
                                                  Style style, 
                                                  boolean verbose)

is intended to return a list of voicings paired with extensions, from which a selection is made.

However, when this section was added:

    /*Init Dan's Classes*/
    if(style.hasCustomVoicing()){

it didn't return a list with the voicing directly. Instead it installed the voicing in the chord, then continued on with the code for dealing with standard voicings, as shown here:

        chord.setVoicing(buffer.toPolylist());
        VoicingDebug.println("Voicing: " + buffer.toPolylist().toString());
        VoicingDebug.println("----------");
        VoicingDebug.println("");
    }
/*End*/
//--------------------------------------

It is kind of jumping the gun, and the code for assessing the list of voicings which calls this will still be executed.

This means that the standard voicing analysis is done on top of the custom, but the voicing is already set in the chord, so the standard analysis probably has no effect.

I think, but am not totally sure, that the reason it does not return to standard after using custom once the leadsheet is played is that the voicings are already stored in the chord, so the selection of a new voicing is ignored.

The simplest fix that restores order, I believe, is to return the list of the single custom voicing and an empty extension list, in the case of the custom code, and let the caller of this method do the actual chord setting, as illustrated below.

        Polylist customVoicing = buffer.toPolylist();
        //chord.setVoicing(buffer.toPolylist());
        VoicingDebug.println("Voicing: " + customVoicing);
        VoicingDebug.println("----------");
        VoicingDebug.println("");
        
        // Return a list of a single voicing with an empty extension
        
        Polylist extension = Polylist.nil;
        return Polylist.list(Polylist.list(customVoicing, extension));
    }
/*End*/


I'm going to make this change for now. If you see a reason not to, we can revert it tomorrow.