Adjusted layout of Rectify CheckBoxes in LickgenFrame for better visibility. Changed the title to Grammar Controls.
Noting that "Continuous" is not the same as Recurrent. Continuous means to generate graded licks continuously. As such, this CheckBox really belongs with Lick Saving and Grading. Size of Selection also belongs there. I tried to move them, but couldn't quite get the layout to work in the time available, so left them as is.