Now passive trading works properly even if the trading quantum does not divide the number of bars evenly.
There is one exception, not due to Grammar.java however: When the quantum is greater than the number of bars in the chorus and user is first, the chorus does not erase as expected. Instead previous contents remain. The Grammar Controls (formerly Lickgen) window shows that the proper rest is being generated, but it is not being installed. Therefore this problem is in some other area, probably LickGen or LickgenFrame.