Checkpoint. CountIn was not working correctly when Impro-Visor starts active trading. The count would be heard on top of the melody.
For now at least, setting Impro-Visor first turns the countIn checkbox
off. This is not so bad, because the user can hear the tempo when 
Impro-Visor plays the first trade.

Something is wrong with passive trading and straight improvisation
however. Lines of notes are getting deleted for no obvious reason.
Have to check whether this has anything to do with recent changes.