Committing first draft of new dual-resolution quantization.
Currently a Quantize Chorus window will open when Impro-Visor starts.
However, this will eventually go away and be openable from within
Utilities (it is now as well).

The old quantization is still there, but this will go away eventually.

The Swing Eighth-Notes checkbox is not yet implemented.

Currently there is an issue with rests being left unnecessarily. Notes
are aligned properly, but rests that could be filled in without damaging
aesthetics are not.

A report of quanta, gcd, notes lost, etc. appears in System.out. This
should be regarded as debugging and eventually removed.