New version of QuantizationDialog.java This solves various ergonomic problems in the previous version.
The ergonomics for notes and triplets are "thermometer" style, while
the ergonomics for absorption will cancel all absorption if any box
is unchecked. (This could be changed to thermometer style in the future
if it is viewd as too different from the rest of the interface.
 
There are also two fields that display the quanta in slots.
These fields can also be edited, even to values other than
those provided by the checkboxes, in case anyone wants to try
some non-standard quanta, or for debugging.