New melodyPart quantization algorithm installed and mostly working. Handles dual quanta and swing. Preserves accidentals in original melody.
There is one case of swing not handled properly yet, which is when
the second third of a swing tuple is tied over to another note. 
In this case the output will be off because some notes will have
too small a duration. Will revisit when fresher.

The old algorithm is still available in the GUI for comparison. There
is a checkbox that causes it to be used instead. This will go away
eventually.

Also, this does not pick the longest candidate competing for a given 
slot as "best". It always uses the first one. This too can be revisited.

Set default quanta to 32nd note and 32nd triplet, as these show up
fine on the leadsheet.