Checkpoint on static quantization. Now swing conversion is correct, except the case where the note or rest in the 1/3 of a beat carries over into the next beat. In that case, we do not attempt to convert. Still I think this is the best yet.
Unfortunately, rest absorption is broken in the new algorithm, so this 
will need to be addressed in a subsequent commit. For no, we default
to no rest absorption.

Now the quantization dialog does not open by default. Instead one
needs to go to Utilities > Open Quantization to get it.

The old quantization popup is no longer in notate, as that algorithm
seems to be inferior to present one.