Basic real-time swing is working. Handles simple cases as shown below. Trace is currently turned on.
Development was complicated by the fact that when a note is first
inserted in real-time, it is extended to a longer note then later 
truncated, in the same way as note entry in leadsheet editing.

Another complicating factor is that the duration of a note is not
determined until either a noteOff or another noteOn message is sent.
So determination of swing requires memory of two previous notes, 
not just one.

Another thing that makes debugging difficult is that certain notes
that are swung perceptually will be interpreted as straight eighths
anyway, due to normal quantization.

Example of trace output:

swing detected at beat 3 converting eb4/3 & f1+16+32/3 to eb8 & f8

swing detected at beat 4 converting g4/3 & f1+2+4+16+32/3 to g8 & f8

swing detected at beat 5 converting eb4/3 & d1+2+16+32/3 to eb8 & d8