Update LSTM code and add LSTM-based generation methods to Impro-Visor
This commit updates the LSTM code to the newest version. It also adds classes
to Impro-Visor that use the LSTM code to improvise on top of chords, both
in regular improvisation and in passive trading mode.

To make building easier, this also modifies the build.xml file to automatically
set up ivy and configure the NetBeans project source classpaths.