Add tentative just-in-time mode for LSTM network
Since the network can take a long time to generate solos on long
leadsheets, it may be advantageous to generate the solo in real
time while the leadsheet is playing. This commit adds a mode that
does this, generating in 4-bar chunks.

The implementation uses two main parts: some lazy-generation variables
that allow Notate to insert new versions of licks as they become
available, and a hot-swap mode in the Midi rendering pathway that
ensures that the sequencer plays at the correct point in time.

There is a noticeable lag each time the sequence is updated, likely
due to the stopping, re-rendering, reloading, and starting of the midi
sequence. Hopefully it will be possible to perform the additions
differently so that this does not occur.