Modify hot-swap system to remove just-in-time generation lag
If just-in-time generation is enabled, instead of actually setting a
different sequence, we can edit the existing sequence in-place. Doing
this causes the MIDI sequencer to continue playing without having to
stop and start again, producing a better transition. Some clicking can
be heard occasionally, but it is much improved.

Just-in-time is now the default LSTM generation mode.