Add probability adjustment sliders to LSTM config
As an experiment, this adds two sliders to the LSTM network config
panel, which control how the probablities are sampled from. The first
slider determines the "risk level", i.e. how likely the network should
choose something with low probability. The second determines which
expert's opinion gets the most weight.

Unfortunately, results don't sound super good if either slider is
adjusted more than a tiny amount. It appears that the interval network
is entirely incapable of doing rhythm on its own (which I suppose was
delegated to the chord network during tarining), and the chord network
doesn't always pick the best pitches either. Also, it is common for
"rest" to be the most likely option, so reducing risk level often
causes the network to refrain from playing anything.

It is possible that applying a different level of scaling to the
rest and sustain probabilities might alleviate the problem, but otherwise
this doesn't seem like a particularly useful feature.