Add post-processing modes (rectification, force-play) to LSTM network
After the network produces the probability distribution to sample from,
we can modify the distribution to fit some desired characteristics before
we sample. Using this, we can implement some "cheats" that can modify
how the network output sounds.

Rectification zeroes out the probability to play a note outside the chord,
leaving only the chord (and optionally color) tones. This ensures that
the network only plays notes that fit the chord.

Force-play allows us to force the network to play a note by zeroing the
probability of sustaining or resting. One useful application is counting
the number of consecutive rests generated by the network, and forcing it
to play a note if it has played rests for too long. This prevents long
rests from appearing in the output.

As a variant of force-play mode, we can instead reset the network state
each time it plays rests for too long. This has the disadvantage of being
discontinuous (i.e. the network's state is wiped), but can force the
network out of some undesired behaviors.