Add great moments connectome and connectome training data
Great moments connectome doesn't seem to work very well. Possibly because of
the different styles of the input, possibly because of dataset size.

Training data for each connectome is in the training-data subdirectory.
The training data for the combination connectome is probably correct, but
might be slightly incorrect, since the connectome was trained a long time
ago and the exact training command is not available.
The training data for the combination-phrase connectome is just the data
for the combination connectome, but restricted to pieces longer than 12 bars.