In Active Trading, Chop and Memorize response was not being rectified, resulting in bad sounding stuff some of the time. Now it is rectified.
However, this class should be refactored to emphasize functional
programming, rather than side-effecting an instance variable. It will
make for more flexibility and rational extensions in the future.
One possibility is to have a checkbox for rectification, etc.
Removing very short rests is another thing that could be added to
the pipeline.