Add "Merge repeated pitches" rectification option to LSTM generation
When this option is checked, the generation process will not play the same
pitch in multiple consecutive notes, but will instead just sustain the
previous note.

With other generation modes, this is done as a post-processing step after
the whole melody is generated. However, in this case, it is easier to modify
the probability distribution so that such notes are never chosen in the first
place. This is accomplished by setting the probability of playing the same
pitch to 0, and then adding the original probability of doing so to the
probability of sustaining the previous note.