Add better error handling for connectome loading
Strangely, if you try to use a ZipInputStream to load something that
is not a zip file, it silently fails by treating the file as an empty
zip file. This commit checks for this case and throws an exception if
it happens.

Additionally, if the user tries to improvise using Deep Learning mode
and the connectome isn't properly loaded, it now opens a warning
dialog instead of silently using grammars instead.