Add configure method to Loadable, and implement 3-expert generative model
This adds support in the NetworkConnectomeLoader for configuration files,
and allows implementations of Loadable to perform configuration using
a config.txt file in the connectome zip.

Additionally, based on the contents of the config.txt file, the
generative model now can operate in 3-expert mode, where two experts
are responsible for pitch and a third expert is specialized in rhythm.