I chased down the problem of why Quotes were not showing up in the Advice dialog:
The Preference visible-advice-components needs to be set to 4095 
(= a mask of 11111111111 binary) in order to see all components.
But we never changed this as Bricks and then Rhythms were added, so
the default was 1023 (= a mask of 00111111111) so Quotes were being
blocked.
Bricks were still getting through somehow. 

The non-default value of this preference is determined by checkboxes
in Global Preferences. We need to go in and add checkboxes for Rhythms
and Bricks to get complete this story.