Add files via upload
When the Improvisation button was pressed and the score started playing, there wasn't a guarantee that autoscroll would work (sometimes it worked, sometimes it didn't. I described the bug a bit more here https://sourceforge.net/p/impro-visor/bugs/70/ . Added autoScrollOnPlayback=true on improviseContinously(), so that autoscroll would work when the Improvisation button was pressed.