Now roadmap brick coloration is defined in the preferences file. If it is not present, a default will be installed and saved in the preference file the first time a roadmap is created.
The handling of roadmap-colors is similar to that of transposing-instruments.
Care must be taken: The SECOND element of these preferences items is
a complete list in itself. It is not the case that the REST of the
preference item as a list is used. If this distinction is not followed,
there will be problems.